feat(frontend): ウィンドウ設定ポップアップのUIを刷新
カード型セクション・数値バッジ・スクロール・トグル化。設定内容は不変。
WindowSettingsPanel を分離。order-009 / test-spec-021 を追加。
