+++ Frontend/docs/orders/order-009.md
@@ -1,0 +1,50 @@
+# order-009: ウィンドウ設定ポップアップの UI 改善
+
+## 日付
+
+2026-05-29
+
+## 参加者
+
+- ユーザー（担当者）
+- Codex
+
+---
+
+## 指示要約
+
+各ウィンドウヘッダーから開く **設定ポップアップ** が見づらいため、設定項目・ストア・保存値は変更せず、ユーザーが操作しやすい UI デザインに刷新する。
+
+---
+
+## 実装方針
+
+1. **`WindowSettingsPanel` をコンポーネント化**
+   - `LayoutEngine.tsx` 内の肥大化していたパネル実装を `presentation/components/WindowSettingsPanel.tsx` へ移す。
+   - ロジック（Zustand ストア、スライダー範囲、マイク選択、モード切替）は従来どおり。
+
+2. **パネル全体**
+   - 幅を `320px` まで拡大、角丸・アクセントのトップバー・グロー付きシャドウでウィンドウから浮き上がる見た目に。
+   - ヘッダーを固定し、本文は `max-height` + 縦スクロールで長い設定でもはみ出さない。
+
+3. **ヘッダー**
+   - アイコンをアクセント枠付きバッジに、タイトル + サブタイトル「表示の調整」で文脈を明示。
+   - 閉じるボタンを 32px タップ領域に拡大。
+
+4. **セクション**
+   - `SettingsSection` でカード型にグルーピング（大文字の英字ラベル tracking を廃止し日本語ラベルを読みやすく）。
+   - 各スライダーを個別の行カードにし、右端にアクセント色の数値バッジを表示。
+
+5. **自動切り替え（バブル）**
+   - ON/OFF の全幅ボタンを、説明文付きの **トグルスイッチ** UI に変更（挙動は `setAutoSwitchEnabled` のまま）。
+
+6. **スコープ外**
+   - 設定値の追加・削除、閾値変更、グローバル設定モーダル（`SettingsModal`）は対象外。
+
+---
+
+## 完了条件
+
+- 文字起こし・バブル・ランキング・説明の各設定が従来どおり変更・保存できること
+- ポップアップがスクロール可能で、視認性が向上していること
+- `npm run build` が成功すること
