+++ Frontend/docs/tests/test-spec-021.md
@@ -1,0 +1,47 @@
+# test-spec-021: ウィンドウ設定ポップアップ UI 改善の検証
+
+## 日付
+
+2026-05-29
+
+## 対象ファイル
+
+- `Frontend/src/presentation/components/WindowSettingsPanel.tsx`
+- `Frontend/src/presentation/layout/LayoutEngine.tsx`
+
+## テストの目的
+
+order-009 に従い、設定ポップアップの見た目と操作性が改善され、設定内容・永続化が破綻しないことを確認する。
+
+## テスト項目
+
+| # | テストケース | 種別 | 期待結果 |
+|---|---|---|---|
+| 1 | パネル表示 | 手動確認 | アクセントトップバー・カード型セクション・数値バッジ付きスライダーが表示される |
+| 2 | スクロール | 手動確認 | バブル設定など項目が多い場合、本文のみ縦スクロールできる |
+| 3 | 文字起こし設定 | 手動確認 | マスター/通常/重要単語スライダーが動作し再読込後も保持される |
+| 4 | バブル設定 | 手動確認 | サイズ系スライダーと自動切り替えトグル・間隔スライダーが動作する |
+| 5 | ランキング設定 | 手動確認 | 要素サイズ/フォント/表示単語数が動作する |
+| 6 | 説明設定 | 手動確認 | 詳細テキストフォントサイズが動作する |
+| 7 | 閉じる操作 | 手動確認 | ×・外側クリック・再度歯車で開閉できる |
+| 8 | 回帰（ビルド） | ビルド | `npm run build` が成功する |
+
+## 手動確認手順
+
+1. 発表中でバブル・文字起こし・ランキング・説明を表示する。
+2. 各ウィンドウの設定を開き、レイアウト（カード・バッジ・トグル）を確認する。
+3. スライダーとバブルの自動切り替えを操作し、表示が即時反映されることを確認する。
+4. ページをリロードし、設定が `localStorage` から復元されることを確認する。
+
+## 結果
+
+| # | 結果 | 備考 |
+|---|---|---|
+| 1 | 未実施 | |
+| 2 | 未実施 | |
+| 3 | 未実施 | |
+| 4 | 未実施 | |
+| 5 | 未実施 | |
+| 6 | 未実施 | |
+| 7 | 未実施 | |
+| 8 | 未実施 | |
