+++ Frontend/src/presentation/layout/LayoutEngine.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react'
-import { GripHorizontal, RefreshCw, Settings, SlidersHorizontal, X } from 'lucide-react'
+import { GripHorizontal, Settings, X } from 'lucide-react'
 import type { DropZone, LayoutNode } from '../../domain/entities/Layout'
 import { movePanel, updateRatio } from './layoutUtils'
 import { getWindowDefinition } from '../windows/registry'
@@ -11,45 +11,7 @@ import {
   SYSTEM_CONTROL_DOCK_MIN_WIDTH_PX,
 } from '../constants/systemControlWindow'
 import { getAccentRgb } from '../../theme/accentTokens'
-import { useTranscription } from '../hooks/useTranscription'
-import { accentRgba } from '../../theme/accentStyles'
-import type { TranscriptionMode } from '../../domain/interfaces/ITranscriptionService'
-import {
-  TRANSCRIPTION_IMPORTANT_FONT_SIZE_MAX,
-  TRANSCRIPTION_IMPORTANT_FONT_SIZE_MIN,
-  TRANSCRIPTION_MASTER_FONT_SCALE_MAX,
-  TRANSCRIPTION_MASTER_FONT_SCALE_MIN,
-  TRANSCRIPTION_PLAIN_FONT_SIZE_MAX,
-  TRANSCRIPTION_PLAIN_FONT_SIZE_MIN,
-  useTranscriptionWindowSettingsStore,
-} from '../../stores/transcriptionWindowSettingsStore'
-import {
-  TERM_MAP_AUTO_SWITCH_INTERVAL_MAX,
-  TERM_MAP_AUTO_SWITCH_INTERVAL_MIN,
-  TERM_MAP_BUBBLE_SIZE_SCALE_MAX,
-  TERM_MAP_BUBBLE_SIZE_SCALE_MIN,
-  TERM_MAP_MAX_VISIBLE_TERMS_MAX,
-  TERM_MAP_MAX_VISIBLE_TERMS_MIN,
-  TERM_MAP_MASTER_SIZE_SCALE_MAX,
-  TERM_MAP_MASTER_SIZE_SCALE_MIN,
-  TERM_MAP_TEXT_FONT_SIZE_MAX,
-  TERM_MAP_TEXT_FONT_SIZE_MIN,
-  useTermMapWindowSettingsStore,
-} from '../../stores/termMapWindowSettingsStore'
-import {
-  IMPORTANCE_RANKING_FONT_SIZE_MAX,
-  IMPORTANCE_RANKING_FONT_SIZE_MIN,
-  IMPORTANCE_RANKING_MASTER_SIZE_SCALE_MAX,
-  IMPORTANCE_RANKING_MASTER_SIZE_SCALE_MIN,
-  IMPORTANCE_RANKING_VISIBLE_COUNT_MAX,
-  IMPORTANCE_RANKING_VISIBLE_COUNT_MIN,
-  useImportanceRankingWindowSettingsStore,
-} from '../../stores/importanceRankingWindowSettingsStore'
-import {
-  DETAIL_WINDOW_FONT_SIZE_MAX,
-  DETAIL_WINDOW_FONT_SIZE_MIN,
-  useDetailWindowSettingsStore,
-} from '../../stores/detailWindowSettingsStore'
+import { WindowSettingsPanel } from '../components/WindowSettingsPanel'
 
 const DropOverlay: React.FC<{ zone: DropZone; rgb: string }> = ({ zone, rgb }) => {
   const base: React.CSSProperties = {
@@ -131,403 +93,6 @@ const systemControlMinStyle = (node: LayoutNode): React.CSSProperties =>
         minHeight: SYSTEM_CONTROL_DOCK_MIN_HEIGHT_PX,
       }
     : { minWidth: 0, minHeight: 0 }
-
-interface SettingsSliderProps {
-  label: string
-  valueLabel: string
-  value: number
-  min: number
-  max: number
-  step: number
-  darkMode: boolean
-  accentRgb: string
-  onChange: (value: number) => void
-}
-
-const SettingsSlider: React.FC<SettingsSliderProps> = ({
-  label,
-  valueLabel,
-  value,
-  min,
-  max,
-  step,
-  darkMode,
-  accentRgb,
-  onChange,
-}) => (
-  <div>
-    <div className="mb-1 flex items-center justify-between gap-2">
-      <span className="text-[11px] font-bold">{label}</span>
-      <span className={`text-[10px] font-mono font-bold ${darkMode ? 'text-slate-400' : 'text-slate-500'}`}>
-        {valueLabel}
-      </span>
-    </div>
-    <input
-      type="range"
-      min={min}
-      max={max}
-      step={step}
-      value={value}
-      onChange={e => onChange(Number(e.target.value))}
-      className="w-full cursor-pointer"
-      style={{ accentColor: `rgb(${accentRgb})` }}
-      aria-label={label}
-    />
-  </div>
-)
-
-interface WindowSettingsPanelProps {
-  windowId: string
-  label: string
-  darkMode: boolean
-  accentRgb: string
-  onClose: () => void
-}
-
-const WindowSettingsPanel: React.FC<WindowSettingsPanelProps> = ({
-  windowId,
-  label,
-  darkMode,
-  accentRgb,
-  onClose,
-}) => {
-  const {
-    mode,
-    setMode,
-    microphones,
-    selectedMicrophoneId,
-    selectMicrophone,
-    refreshMicrophones,
-  } = useTranscription()
-  const masterFontScale = useTranscriptionWindowSettingsStore(s => s.masterFontScale)
-  const plainTextFontSizePx = useTranscriptionWindowSettingsStore(s => s.plainTextFontSizePx)
-  const importantTermFontSizePx = useTranscriptionWindowSettingsStore(s => s.importantTermFontSizePx)
-  const setMasterFontScale = useTranscriptionWindowSettingsStore(s => s.setMasterFontScale)
-  const setPlainTextFontSizePx = useTranscriptionWindowSettingsStore(s => s.setPlainTextFontSizePx)
-  const setImportantTermFontSizePx = useTranscriptionWindowSettingsStore(s => s.setImportantTermFontSizePx)
-  const termMapMasterSizeScale = useTermMapWindowSettingsStore(s => s.masterSizeScale)
-  const termMapBubbleSizeScale = useTermMapWindowSettingsStore(s => s.bubbleSizeScale)
-  const termMapTextFontSizePx = useTermMapWindowSettingsStore(s => s.textFontSizePx)
-  const termMapAutoSwitchEnabled = useTermMapWindowSettingsStore(s => s.autoSwitchEnabled)
-  const termMapAutoSwitchIntervalSec = useTermMapWindowSettingsStore(s => s.autoSwitchIntervalSec)
-  const termMapMaxVisibleTerms = useTermMapWindowSettingsStore(s => s.maxVisibleTerms)
-  const setTermMapMasterSizeScale = useTermMapWindowSettingsStore(s => s.setMasterSizeScale)
-  const setTermMapBubbleSizeScale = useTermMapWindowSettingsStore(s => s.setBubbleSizeScale)
-  const setTermMapTextFontSizePx = useTermMapWindowSettingsStore(s => s.setTextFontSizePx)
-  const setTermMapAutoSwitchEnabled = useTermMapWindowSettingsStore(s => s.setAutoSwitchEnabled)
-  const setTermMapAutoSwitchIntervalSec = useTermMapWindowSettingsStore(s => s.setAutoSwitchIntervalSec)
-  const setTermMapMaxVisibleTerms = useTermMapWindowSettingsStore(s => s.setMaxVisibleTerms)
-  const importanceRankingMasterSizeScale = useImportanceRankingWindowSettingsStore(s => s.masterSizeScale)
-  const importanceRankingFontSizePx = useImportanceRankingWindowSettingsStore(s => s.fontSizePx)
-  const importanceRankingVisibleCount = useImportanceRankingWindowSettingsStore(s => s.visibleCount)
-  const setImportanceRankingMasterSizeScale = useImportanceRankingWindowSettingsStore(s => s.setMasterSizeScale)
-  const setImportanceRankingFontSizePx = useImportanceRankingWindowSettingsStore(s => s.setFontSizePx)
-  const setImportanceRankingVisibleCount = useImportanceRankingWindowSettingsStore(s => s.setVisibleCount)
-  const detailFontSizePx = useDetailWindowSettingsStore(s => s.fontSizePx)
-  const setDetailFontSizePx = useDetailWindowSettingsStore(s => s.setFontSizePx)
-  const dk = darkMode
-
-  const modeButton = (value: TranscriptionMode, text: string) => {
-    const active = mode === value
-    return (
-      <button
-        type="button"
-        onClick={() => setMode(value)}
-        className={`flex-1 rounded-md border px-2 py-1.5 text-[11px] font-bold transition-colors ${active
-          ? ''
-          : dk
-            ? 'border-slate-700 text-slate-400 hover:bg-slate-800'
-            : 'border-slate-200 text-slate-500 hover:bg-slate-50'}`}
-        style={active
-          ? {
-              borderColor: `rgba(${accentRgb},0.62)`,
-              backgroundColor: `rgba(${accentRgb},${dk ? 0.2 : 0.1})`,
-              color: `rgba(${accentRgb},${dk ? 0.96 : 0.88})`,
-            }
-          : undefined}
-      >
-        {text}
-      </button>
-    )
-  }
-
-  return (
-    <div
-      data-window-settings-panel="true"
-      className={`absolute right-2 top-10 z-[70] w-[min(92vw,280px)] rounded-xl border p-3 shadow-2xl ${dk
-        ? 'border-slate-700 bg-[#0d0e1a] text-slate-200'
-        : 'border-slate-200 bg-white text-slate-800'}`}
-      role="dialog"
-      aria-label={`${label} の設定`}
-      onClick={e => e.stopPropagation()}
-    >
-      <div className="mb-3 flex items-center justify-between gap-2">
-        <div className="flex min-w-0 items-center gap-2">
-          <Settings size={14} style={{ color: accentRgba(accentRgb, dk ? 0.9 : 0.75) }} />
-          <p className="min-w-0 truncate text-xs font-black">{label} の設定</p>
-        </div>
-        <button
-          type="button"
-          onClick={onClose}
-          className={`shrink-0 rounded-md p-1 transition-colors ${dk ? 'text-slate-500 hover:bg-slate-800 hover:text-white' : 'text-slate-400 hover:bg-slate-100 hover:text-slate-800'}`}
-          aria-label="設定を閉じる"
-        >
-          <X size={14} />
-        </button>
-      </div>
-
-      {windowId === SYSTEM_CONTROL_WINDOW_ID ? (
-        <div className="space-y-3">
-          <section>
-            <div className={`mb-1.5 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <SlidersHorizontal size={11} /> 文字起こしモード
-            </div>
-            <div className="flex gap-2">
-              {modeButton('fast', '速度重視')}
-              {modeButton('accurate', '品質重視')}
-            </div>
-            <p className={`mt-1 text-[10px] leading-snug ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              {mode === 'fast' ? 'リアルタイム性を優先します。' : '停止後の高精度化を優先します。'}
-            </p>
-          </section>
-
-          <section>
-            <div className="mb-1.5 flex items-center justify-between gap-2">
-              <span className={`text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-                マイク
-              </span>
-              <button
-                type="button"
-                onClick={() => refreshMicrophones()}
-                className={`inline-flex items-center gap-1 rounded-md border px-2 py-0.5 text-[10px] font-bold transition-colors ${dk
-                  ? 'border-slate-600 bg-slate-800 text-slate-300 hover:bg-slate-700'
-                  : 'border-slate-300 bg-white text-slate-600 hover:bg-slate-50'}`}
-                title="マイク一覧を再取得"
-              >
-                <RefreshCw size={10} />
-                更新
-              </button>
-            </div>
-            <select
-              value={selectedMicrophoneId}
-              onChange={e => selectMicrophone(e.target.value)}
-              className={`w-full rounded-lg border px-2.5 py-1.5 text-xs ${dk
-                ? 'border-slate-700 bg-slate-800 text-slate-200'
-                : 'border-slate-300 bg-white text-slate-700'}`}
-            >
-              {microphones.length === 0 && <option value="">利用可能マイクなし</option>}
-              {microphones.map(mic => (
-                <option key={mic.deviceId} value={mic.deviceId}>{mic.label}</option>
-              ))}
-            </select>
-          </section>
-        </div>
-      ) : windowId === 'transcription' ? (
-        <div className="space-y-3">
-          <section>
-            <div className={`mb-2 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <SlidersHorizontal size={11} /> フォントサイズ
-            </div>
-            <div className="space-y-2.5">
-              <SettingsSlider
-                label="マスター"
-                valueLabel={`${Math.round(masterFontScale * 100)}%`}
-                value={masterFontScale}
-                min={TRANSCRIPTION_MASTER_FONT_SCALE_MIN}
-                max={TRANSCRIPTION_MASTER_FONT_SCALE_MAX}
-                step={0.05}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setMasterFontScale}
-              />
-              <SettingsSlider
-                label="通常文字"
-                valueLabel={`${plainTextFontSizePx}px`}
-                value={plainTextFontSizePx}
-                min={TRANSCRIPTION_PLAIN_FONT_SIZE_MIN}
-                max={TRANSCRIPTION_PLAIN_FONT_SIZE_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setPlainTextFontSizePx}
-              />
-              <SettingsSlider
-                label="重要単語"
-                valueLabel={`${importantTermFontSizePx}px`}
-                value={importantTermFontSizePx}
-                min={TRANSCRIPTION_IMPORTANT_FONT_SIZE_MIN}
-                max={TRANSCRIPTION_IMPORTANT_FONT_SIZE_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setImportantTermFontSizePx}
-              />
-            </div>
-          </section>
-        </div>
-      ) : windowId === 'bubbleCloud' ? (
-        <div className="space-y-3">
-          <section>
-            <div className={`mb-2 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <SlidersHorizontal size={11} /> サイズ
-            </div>
-            <div className="space-y-2.5">
-              <SettingsSlider
-                label="マスター"
-                valueLabel={`${Math.round(termMapMasterSizeScale * 100)}%`}
-                value={termMapMasterSizeScale}
-                min={TERM_MAP_MASTER_SIZE_SCALE_MIN}
-                max={TERM_MAP_MASTER_SIZE_SCALE_MAX}
-                step={0.05}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setTermMapMasterSizeScale}
-              />
-              <SettingsSlider
-                label="バブル"
-                valueLabel={`${Math.round(termMapBubbleSizeScale * 100)}%`}
-                value={termMapBubbleSizeScale}
-                min={TERM_MAP_BUBBLE_SIZE_SCALE_MIN}
-                max={TERM_MAP_BUBBLE_SIZE_SCALE_MAX}
-                step={0.1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setTermMapBubbleSizeScale}
-              />
-              <SettingsSlider
-                label="テキスト"
-                valueLabel={`${termMapTextFontSizePx}px`}
-                value={termMapTextFontSizePx}
-                min={TERM_MAP_TEXT_FONT_SIZE_MIN}
-                max={TERM_MAP_TEXT_FONT_SIZE_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setTermMapTextFontSizePx}
-              />
-              <SettingsSlider
-                label="最大表示数"
-                valueLabel={`${termMapMaxVisibleTerms}語`}
-                value={termMapMaxVisibleTerms}
-                min={TERM_MAP_MAX_VISIBLE_TERMS_MIN}
-                max={TERM_MAP_MAX_VISIBLE_TERMS_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setTermMapMaxVisibleTerms}
-              />
-            </div>
-          </section>
-
-          <section>
-            <div className={`mb-2 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <RefreshCw size={11} /> 自動切り替え
-            </div>
-            <button
-              type="button"
-              onClick={() => setTermMapAutoSwitchEnabled(!termMapAutoSwitchEnabled)}
-              className={`mb-2 flex w-full items-center justify-between rounded-lg border px-2.5 py-2 text-xs font-bold transition-colors ${termMapAutoSwitchEnabled
-                ? ''
-                : dk
-                  ? 'border-slate-700 bg-slate-800/50 text-slate-400 hover:bg-slate-800'
-                  : 'border-slate-200 bg-slate-50 text-slate-500 hover:bg-slate-100'}`}
-              style={termMapAutoSwitchEnabled
-                ? {
-                    borderColor: `rgba(${accentRgb},0.62)`,
-                    backgroundColor: `rgba(${accentRgb},${dk ? 0.2 : 0.1})`,
-                    color: `rgba(${accentRgb},${dk ? 0.96 : 0.88})`,
-                  }
-                : undefined}
-              aria-pressed={termMapAutoSwitchEnabled}
-            >
-              <span>{termMapAutoSwitchEnabled ? 'ON' : 'OFF'}</span>
-              <span className="text-[10px] opacity-80">
-                {termMapAutoSwitchEnabled ? '説明と用語を切り替えます' : '手動表示にします'}
-              </span>
-            </button>
-            <SettingsSlider
-              label="切り替え間隔"
-              valueLabel={`${termMapAutoSwitchIntervalSec}秒`}
-              value={termMapAutoSwitchIntervalSec}
-              min={TERM_MAP_AUTO_SWITCH_INTERVAL_MIN}
-              max={TERM_MAP_AUTO_SWITCH_INTERVAL_MAX}
-              step={1}
-              darkMode={dk}
-              accentRgb={accentRgb}
-              onChange={setTermMapAutoSwitchIntervalSec}
-            />
-          </section>
-        </div>
-      ) : windowId === 'importanceRanking' ? (
-        <div className="space-y-3">
-          <section>
-            <div className={`mb-2 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <SlidersHorizontal size={11} /> 表示
-            </div>
-            <div className="space-y-2.5">
-              <SettingsSlider
-                label="要素サイズ"
-                valueLabel={`${Math.round(importanceRankingMasterSizeScale * 100)}%`}
-                value={importanceRankingMasterSizeScale}
-                min={IMPORTANCE_RANKING_MASTER_SIZE_SCALE_MIN}
-                max={IMPORTANCE_RANKING_MASTER_SIZE_SCALE_MAX}
-                step={0.05}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setImportanceRankingMasterSizeScale}
-              />
-              <SettingsSlider
-                label="フォント"
-                valueLabel={`${importanceRankingFontSizePx}px`}
-                value={importanceRankingFontSizePx}
-                min={IMPORTANCE_RANKING_FONT_SIZE_MIN}
-                max={IMPORTANCE_RANKING_FONT_SIZE_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setImportanceRankingFontSizePx}
-              />
-              <SettingsSlider
-                label="表示単語数"
-                valueLabel={`${importanceRankingVisibleCount}個`}
-                value={importanceRankingVisibleCount}
-                min={IMPORTANCE_RANKING_VISIBLE_COUNT_MIN}
-                max={IMPORTANCE_RANKING_VISIBLE_COUNT_MAX}
-                step={1}
-                darkMode={dk}
-                accentRgb={accentRgb}
-                onChange={setImportanceRankingVisibleCount}
-              />
-            </div>
-          </section>
-        </div>
-      ) : windowId === 'detail' ? (
-        <div className="space-y-3">
-          <section>
-            <div className={`mb-2 flex items-center gap-1.5 text-[10px] font-bold uppercase tracking-widest ${dk ? 'text-slate-500' : 'text-slate-400'}`}>
-              <SlidersHorizontal size={11} /> フォントサイズ
-            </div>
-            <SettingsSlider
-              label="詳細テキスト"
-              valueLabel={`${detailFontSizePx}px`}
-              value={detailFontSizePx}
-              min={DETAIL_WINDOW_FONT_SIZE_MIN}
-              max={DETAIL_WINDOW_FONT_SIZE_MAX}
-              step={1}
-              darkMode={dk}
-              accentRgb={accentRgb}
-              onChange={setDetailFontSizePx}
-            />
-          </section>
-        </div>
-      ) : (
-        <p className={`text-xs leading-relaxed ${dk ? 'text-slate-400' : 'text-slate-500'}`}>
-          このウィンドウ固有の設定は今後追加予定です。
-        </p>
-      )}
-    </div>
-  )
-}
 
 export interface LayoutEngineProps {
   layout: LayoutNode
